third_party/WebKit/LayoutTests/http/tests/security/video-poster-cross-origin-crash-expected.txt

CONSOLE ERROR: Access to Image at 'http://localhost:8080/misc/resources/compass.jpg' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
>>>
